Move error tests to separate file
